iris: refactor bootstrap processor initialization
This is done to make it easy to initialize the alternative processors.
