+++ iris/arch/x86/amd64/CMakeLists.txt
@@ -17,17 +17,19 @@ target_sources(
     iris
     PRIVATE arch/x86/amd64/boot/cxx_init.cpp
             arch/x86/amd64/boot/init.cpp
+            arch/x86/amd64/core/print.cpp
             arch/x86/amd64/core/task.cpp
-            arch/x86/amd64/hw/local_apic.cpp
+            arch/x86/amd64/hw/gdt.cpp
+            arch/x86/amd64/hw/idt.cpp
             arch/x86/amd64/hw/io_apic.cpp
             arch/x86/amd64/hw/irq_controller.cpp
             arch/x86/amd64/hw/irq.cpp
-            arch/x86/amd64/hw/idt.cpp
+            arch/x86/amd64/hw/local_apic.cpp
             arch/x86/amd64/hw/pic.cpp
             arch/x86/amd64/hw/power.cpp
             arch/x86/amd64/hw/processor_info.cpp
             arch/x86/amd64/hw/serial.cpp
+            arch/x86/amd64/hw/tss.cpp
             arch/x86/amd64/mm/address_space.cpp
             arch/x86/amd64/mm/map_physical_address.cpp
-            arch/x86/core/print.cpp
 )
