+++ iris/include/iris/arch/x86/amd64/core/processor.h
@@ -1,6 +1,8 @@
 #pragma once
 
 #include <iris/arch/x86/amd64/hw/local_apic.h>
+#include <iris/arch/x86/amd64/segment_descriptor.h>
+#include <iris/arch/x86/amd64/tss.h>
 
 namespace iris {
 class Processor;
@@ -15,8 +17,17 @@ namespace arch {
         x86::amd64::LocalApic& local_apic() { return *m_local_apic; }
         void set_local_apic(x86::amd64::LocalApic local_apic) { m_local_apic = local_apic; }
 
+        auto& gdt() { return m_gdt; }
+        auto& tss() { return m_tss; }
+
+        void set_fallback_kernel_stack(uptr stack) { m_fallback_kernel_stack = stack; }
+        uptr fallback_kernel_stack() const { return m_fallback_kernel_stack; }
+
     private:
         di::Optional<x86::amd64::LocalApic> m_local_apic;
+        di::Array<iris::x86::amd64::sd::SegmentDescriptor, 11> m_gdt {};
+        iris::x86::amd64::TSS m_tss {};
+        uptr m_fallback_kernel_stack {};
     };
 }
