+++ iris/include/iris/arch/x86/amd64/gdt.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <iris/arch/x86/amd64/segment_descriptor.h>
+
+namespace iris::x86::amd64 {
+void init_gdt();
+}
