+++ iris/include/iris/fs/debug_file.h
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <iris/core/interruptible_spinlock.h>
+#include <iris/core/print.h>
 #include <iris/fs/file.h>
 
 namespace iris {
 struct DebugFile {
 private:
-    friend Expected<usize> tag_invoke(di::Tag<read_file>, DebugFile& self, di::Span<di::Byte> data);
+    friend Expected<usize> tag_invoke(di::Tag<read_file>, DebugFile&, di::Span<di::Byte> data);
 
     friend Expected<usize> tag_invoke(di::Tag<write_file>, DebugFile& self, di::Span<di::Byte const> data) {
         auto guard = di::ScopedLock(self.m_lock);
