+++ iris/include/iris/hw/acpi/acpi.h
@@ -5,7 +5,7 @@
 namespace iris::acpi {
 struct AcpiInformation {
     di::Optional<MADT> madt;
-    di::Optional<ProcessorLocalApicStructure> local_apic;
+    di::Vector<ProcessorLocalApicStructure> local_apic;
     di::Vector<IoApicStructure> io_apics;
     di::Vector<InterruptSourceOverrideStructure> interrupt_source_overrides;
     di::Optional<LocalApicNmiStructure> local_apic_nmi;
